+++ packages/dnd/src/isHotkeyPressed.ts
@@ -5,6 +5,8 @@
  * https://github.com/JohannesKlauss/react-hotkeys-hook/blob/bc55a281f1d212d09de786aeb5cd236c58d9531d/src/parseHotkey.ts
  */
 
+import { lc } from 'react-querybuilder';
+
 type ModifierKey = 'shift' | 'alt' | 'meta' | 'mod' | 'ctrl';
 
 // #region parseHotkey.ts
@@ -33,10 +35,7 @@ const mappedKeys: Record<string, string> = {
 };
 
 const mapKey = (key?: string) =>
-  ((key && mappedKeys[key]) || key || '')
-    .trim()
-    .toLowerCase()
-    .replace(/key|digit|numpad|arrow/, '');
+  lc(((key && mappedKeys[key]) || key || '').trim()).replace(/key|digit|numpad|arrow/, '');
 
 const isHotkeyModifier = (key: string) => reservedModifierKeywords.has(key as ModifierKey);
 // #endregion parseHotkey.ts
@@ -90,7 +89,7 @@ const isReadonlyArray = (value: unknown): value is readonly unknown[] => Array.i
 
 export const isHotkeyPressed = (key: string | readonly string[], splitKey = ','): boolean =>
   (isReadonlyArray(key) ? key : key.split(splitKey)).every(hotkey => {
-    const hk = hotkey.trim().toLowerCase();
+    const hk = lc(hotkey.trim());
     return currentlyPressedKeys.has(keyAliases[hk] ?? hk);
   });
 
@@ -105,12 +104,12 @@ const pushToCurrentlyPressedKeys = (key: string | string[]) => {
   if (currentlyPressedKeys.has('meta')) {
     for (const key of currentlyPressedKeys) {
       if (!isHotkeyModifier(key)) {
-        currentlyPressedKeys.delete(key.toLowerCase());
+        currentlyPressedKeys.delete(lc(key));
       }
     }
   }
 
-  for (const hotkey of hotkeyArray) currentlyPressedKeys.add(hotkey.toLowerCase());
+  for (const hotkey of hotkeyArray) currentlyPressedKeys.add(lc(hotkey));
 };
 
 const removeFromCurrentlyPressedKeys = (key: string | string[]) => {
@@ -124,7 +123,7 @@ const removeFromCurrentlyPressedKeys = (key: string | string[]) => {
   if (key === 'meta') {
     currentlyPressedKeys.clear();
   } else {
-    for (const hotkey of hotkeyArray) currentlyPressedKeys.delete(hotkey.toLowerCase());
+    for (const hotkey of hotkeyArray) currentlyPressedKeys.delete(lc(hotkey));
   }
 };
 // #endregion isHotkeyPressed.ts